- fix windows compilation. Sqrt require double or float (not int).

+++ src/tools/points_detector.cc
@@ -137,7 +137,7 @@ void SaveDescriptorAsPatches( const libmv::vector<descriptor::Descriptor *> & de
   for (int i = 0; i < desc.size(); ++i) {
     const descriptor::VecfDescriptor * array = dynamic_cast<descriptor::VecfDescriptor *>( desc[i] );
     if(array) {
-      const int size = sqrt(array->coords.size());
+      const int size = sqrt((double)(array->coords.size()));
       Array3Df ima(size,size);
       for(int j=0; j< size; ++j)
       for(int i=0; i< size; ++i)
